requirements.txt
tensorflow==2.10.0
streamlit==1.14.0
Pillow==9.1.0
numpy==1.21.4
matplotlib==3.5.0
